fix(#112): collapse three orchestrators into StreamingEngine; delete pipeline.py + orchestrator.py
PipelineOrchestrator was instantiated in StreamingEngine.__init__ but never
called — all multi-stream orchestration already lived in stream_data().
Pipeline was a thin config-translator wrapper that the runner was the only
caller of.

- Delete src/engine/orchestrator.py and src/engine/pipeline.py outright.
- Move the contract→connector config translation helpers (_build_config_dict,
  _translate_source_config, _translate_database_source, _translate_api_source,
  _build_destination_config, _translate_assignment) to src/runner.py, which is
  now the single caller.
- PipelineRunner.run() creates StreamingEngine directly and calls
  engine.stream_data(config_dict); Pipeline is no longer needed.
- Remove the dead self.orchestrator assignment and PipelineOrchestrator import
  from StreamingEngine.__init__.
- Remove Pipeline from src/__init__ exports.
- Delete test_orchestrator.py; remove TestDeepMergeDicts from test_engine_unit.py
  (both tested deleted code); update test_package_init.py and
  test_subpackage_inits.py to reflect the removed Pipeline export; repoint
  test_nested_assignments.py's _translate_assignment import to src.runner.
