+++ src/engine/engine.py
@@ -23,7 +23,6 @@ from .exceptions import (
     ConfigurationError, StreamProcessingError, StreamExecutionError,
     StreamConfigurationError, StageConfigurationError
 )
-from .orchestrator import PipelineOrchestrator
 
 # gRPC imports for destination streaming
 from ..grpc.client import DestinationGRPCClient, BatchResult, generate_record_id
@@ -67,9 +66,6 @@ class StreamingEngine:
         
         # Setup structured logging
         self.logger = logging.getLogger(f"{__name__}.{pipeline_id}")
-        
-        # Pipeline orchestration
-        self.orchestrator = PipelineOrchestrator(pipeline_id)
 
         # Fault tolerance components with state management
         self.sharded_state_manager = StateManager(pipeline_id, "./state")
