+++ tests/unit/analitiq_stream/test_package_init.py
@@ -25,7 +25,6 @@ class TestMainPackageInit:
 
         expected_exports = [
             "StreamingEngine",
-            "Pipeline",
             "RetryHandler",
             "CircuitBreaker",
             "DeadLetterQueue",
@@ -44,14 +43,12 @@ class TestMainPackageInit:
             BaseConnector,
             CircuitBreaker,
             DeadLetterQueue,
-            Pipeline,
             RetryHandler,
             StreamingEngine,
         )
 
         for cls in (
             StreamingEngine,
-            Pipeline,
             RetryHandler,
             CircuitBreaker,
             DeadLetterQueue,
@@ -105,15 +102,12 @@ class TestImportErrors:
     def test_circular_import_prevention(self):
         """Test that imports don't create circular dependencies."""
         # Import in different orders to test for circular imports
-        # Use new module paths (src.engine, src.source.connectors)
-        import src.engine.pipeline
         import src.engine.engine
         import src.source.connectors.api
         import src
 
         # All imports should succeed
         assert all([
-            src.engine.pipeline,
             src.engine.engine,
             src.source.connectors.api,
             src
